refactor: update parameter names for consistency and clarity
