+++ wind_forecasting/tuning/utils/path_utils.py
@@ -1,30 +1,9 @@
 """
-Path resolution and dictionary utilities.
+Dictionary utilities for tuning operations.
 """
 import collections.abc
-from pathlib import Path
-from typing import Dict, Optional, Union
-
-
-def resolve_path(base_path: str, path_input: Optional[Union[str, Path]]) -> Optional[str]:
-    """
-    Resolve a path relative to a base path or return absolute path.
-    
-    Args:
-        base_path: Base path for relative path resolution
-        path_input: Path to resolve (can be relative or absolute)
-        
-    Returns:
-        Resolved absolute path as string, or None if path_input is None
-    """
-    if not path_input:
-        return None
-    # Convert potential Path object back to string if needed
-    path_str = str(path_input)
-    abs_path = Path(path_str)
-    if not abs_path.is_absolute():
-        abs_path = Path(base_path) / abs_path
-    return str(abs_path.resolve())
+from typing import Dict
+from wind_forecasting.utils.path_utils import resolve_path
 
 
 def flatten_dict(d: Dict, parent_key: str = '', sep: str = '.') -> Dict:
